Rectifying some compiler warnings

+++ gc_memory/memory.c
@@ -3125,7 +3125,7 @@ void read_rom()
    else
      //*rdword = *((unsigned long *)(rom + (address & 0x03FFFFFF)));
      assert(rdword);
-     ROMCache_read((unsigned int*)rdword+1, (address & 0x03FFFFFF), 4);
+     ROMCache_read((u8*)rdword+1, (address & 0x03FFFFFF), 4);
      sign_extended(*rdword);
 }
 
@@ -3133,7 +3133,7 @@ void read_romb()
 {
    //*rdword = *(rom + ((address^S8) & 0x03FFFFFF));
    assert(rdword);
-   ROMCache_read((unsigned int*)((char*)rdword+7), (address & 0x03FFFFFF), 1);
+   ROMCache_read((u8*)((char*)rdword+7), (address & 0x03FFFFFF), 1);
    sign_extendedb(*rdword);
 }
 
@@ -3141,7 +3141,7 @@ void read_romh()
 {
    //*rdword = *((unsigned short *)(rom + ((address^S16) & 0x03FFFFFF)));
    assert(rdword);
-   ROMCache_read((unsigned int*)((short*)rdword+3), (address & 0x03FFFFFF), 2);
+   ROMCache_read((u8*)((short*)rdword+3), (address & 0x03FFFFFF), 2);
    sign_extendedh(*rdword);
 }
 
@@ -3150,7 +3150,7 @@ void read_romd()
    //*rdword = ((unsigned long long)(*((unsigned long *)(rom+(address&0x03FFFFFF))))<<32)|
    //  *((unsigned long *)(rom + ((address+4)&0x03FFFFFF)));
    assert(rdword);
-   ROMCache_read((unsigned int*)rdword, (address & 0x03FFFFFF), 8);
+   ROMCache_read((u8*)rdword, (address & 0x03FFFFFF), 8);
 }
 
 void write_rom()
